+++ core/src/main/java/nl/weeaboo/vn/gdx/res/GeneratedResource.java
@@ -4,14 +4,13 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.Serializable;
 
-import com.badlogic.gdx.utils.Disposable;
-
 import nl.weeaboo.common.Checks;
 import nl.weeaboo.io.CustomSerializable;
+import nl.weeaboo.vn.core.IDestructible;
 import nl.weeaboo.vn.impl.core.StaticRef;
 
 @CustomSerializable
-final class GeneratedResource<T extends Serializable & Disposable> implements IResource<T> {
+final class GeneratedResource<T extends Serializable & IDestructible> implements IResource<T> {
 
     private static final long serialVersionUID = 1L;